Changed Splash Screen and Placeholder Graph Images to match Login Page Dark Mode PR (#564)
## Summary
Changed splash screen and placeholder / thumbnail graph images to match
recent Login Page Dark Mode PR #561, for consistency. Fixes new issue
#563.

Subjectively, this seems to be better to match the other pending dark
mode PR, #561. This also changes the placeholder image while waiting for
an internet connection to load a graph, but due to being dark, your eye
kind of glosses over it. Couldn't get a screenshot of placeholder image
while loading graph because it went by so fast.

## Validation
App launches on device (iOS 18.2 iPhone 13 Pro) and in simulator. (iOS
18.2 iPhone 16 Pro) There are no code changes, merely .png asset swaps
with no changes to the identifying .json files, and the new PNGs are the
same dimensions as the old ones.