Add a CODEOWNERS file (#570)
## Summary
It's becoming a little hard for me to keep track of PR states. To help
with this, I'm going to experiment with configuring github to request a
review from me on all PRs. Then I can filter to PRs waiting on review
from me to know which need my attention.

CODEOWNERS appears the simplest way to do this.

## Validation
As this is a git change it's hard to verify pre-merge. Will merge, and
then check PRs start listing me as a reviewer.