Deprecate storing user data in UserDefaults (#574)
## Summary
Previously we were double-writing to CoreData and to UserDefaults in
case we decided not to move ahead with CoreData. Now we have a public
release that is fully migrated to CoreData we are committed to the
migration, and don't need to maintain the UserDefaults paths.

This leaves in the code which migrates data from UserDefaults to
CoreData, so it it still possible to seamlessly upgrade from previous
versions, but removes incremental updates to UserDefaults, and deletes
all values after the data has been migrated. This means downgrading from
this version to <6.7 will require logging in again.

## Validation
* [x] Launch the app and verify you can reload goals, view timezone
* [x] Verify it is still possible to seamlessly upgrade from code before
the CoreData migration