Switch the gallery view to use a StackView instead of manual layout (#542)
## Summary
Previously we used SnapKit and lots of constraints in order to control
the layout of the gallery view. This was a bunch of code, and cause
constraint violation warnings due to margins being impossible in
conjunction with hiding elements by setting their height to 0.

Here we migrate to use UIStackView which automatically handles
constraints, sizing, and removing hidden views. This lets us delete a
lot of code, and also fixes a bug where there would be a rendering
glitch if returning to the gallery view from a goal view where the
keyboard was hidden - previously the area the keyboard had occupied
would appear blank.

Known issues to fix before merging: The background for last updated,
deadbeat, etc, do not use the full width of the screen.

## Validation
Loaded the app on iPhone and iPad
* Checked the layout looks good in portrait and landscape
* Checked with and without keyboard
* Checked navigating from goal view back to gallery view
* Confirmed with the keyboard visible you can still scroll all goals