Remove ServiceLocator references from GalleryViewController (#575)
## Summary
I'd like to move the app away from having direct references to
ServiceLocator and instead have classes gain references to the
dependencies they need via dependency injection.

This makes a start in that direction with GalleryViewController. Now
this is created in code via SceneDelegate we can pass in its
dependencies at construction time.

## Validation
* [x] Verify that the app can launch and show the gallery view
* [x] Verify it is possible to pull to refresh the goals list