Do not show deadbeat warning before loading user (#587)
## Summary
App started off claiming user is a deadbeat.
The state resulted from recent changes in #583 where the deadbeat view
was hidden or not only when a user was logged in _and_ the default state
being to show the deadbeat view.


## Validation
signed out, started app anew, signed in
in simulator


Fixes #586