Pull to refresh user on settings page (#594)
## Summary
Settings shows the user's timezone but only a few places in the app were
fetching the user: fetching default notification settings (entering the
default notification settings screen) and upon signing in.
This MR allows the user to trigger a fetch of the User endpoint via a
Pull-To-Refresh on the Settings screen.

## Validation
used the app in the simulator
change the timezone at the website
tapped around in the app again both entering settings and pull to
refresh from settings


Fixes #507 and helps out with #360