Fix CurrentUserManager nonisolated warnings (#584)
## Summary
CurrentUserManager has some methods which need to be called outside the
actor - a mix of methods and constructor. These used some properties
which then had to be nonisolated, producing lots of warnings. None of
these objects are actually stateful, so just create them on demand in
the methods that use them.

## Validation
* [ ] Launch the app
* [ ] Verify it is possible to log out and log in