bump Gallery's collectionView to UICollectionViewDiffableDataSource (#614)
## Summary
Migrated the gallery's collection view to the more modern api, using
diffable datasource and snapshots. See
https://github.com/beeminder/BeeSwift/pull/607#discussion_r1904915913.

Transitions from one state to another (one set of goals to the next) is
animated.

## Validation
ran app in simulator, both iPhone and iPad (easier to see multiple
columns and animation, for example, when filtering)
rotated simulator
changed sort preference
filtered on gallery's goals
signed out
signed in
navigated from gallery to goal and back
assured that some goal's data changed (for example safeBuf and safeSum)
and the effect was seen in the cell at the same position