Gallery's list is now fed by nsfetchedresultscontroller (#607)
## Summary
The initial loading (displaying) of goals on the gallery was slow. The
app was still relying on passing and receiving
nsnotification/notifications.
With the data in Core Data, the NSFetchedResultsController can be used.

Migrated galleryVC to use the fetched results controller.
Updated the sorting implementation accordingly, directly applying the
sort type to the fetchRequest itself.
Updated the filtering as well to also apply directly to the
fetchRequest.


## Validation
Ran the app in the simulator.
Changed sorting.
Changed filters.
Signed in.
Signed out.
Navigated from Gallery to Goal and back.
Sent app into background. Brought back into foreground.
Force quit app, relaunched app.


## Tickets
helps with #588