Do not commit build numbers to git (#581)
## Summary
Each testflight build has a unique build number. This is required by
apple, and also helps identify e.g. which builds crashes are present in.
Previously every time we did a build we would update this and check in
the new value. This was creating a medium amount of commit noise, so
instead we'll still update the value, but not check in the changes.

See #571 for more discussion

## Validation
Ran `fastlane ios beta`
Observed
* A build was created and uploaded with the appropriate build number
* No commits are created
* The repo is clean after the command completes.