calculate the deadline correctly from the picker
The same calculation was being used for computing
alertStart and deadline offsets. Deadline however
can be - or +, - for times between 07:00 and 23:59
and positive for times between 00:01 and 06:00.
This resulted in invalid values being sent from
app to the server for many of the deadline possibilities
and thus the backend replying with an error code.