Disable animation for goal thumbnails (#625)
## Summary
Animation of goal thumbnails in the gallery often had some unfortunate
affects:
* Because gallery cells are re-used, on returning to the gallery from a
goal the gallery entry for a goal would show the graph _for a different
goal_ before fading to the right one
* Refreshing the gallery would cause images to flicker to slightly
darker and back

It may be possible to more carefully choose when to animate to address
these, but for now avoid the issue by not animating thumbnails.

## Validation
Loaded the app and navigated to a goal and back. Observed the gallery
immediately showed the correct graphs.