Fix gallery not updating list based on sort (#618)
## Summary
Gallery was not updating how its list was sorted after the preference
had been adjusted. Also, the gallery's last updated was meant to display
when the goals were last fetched.


### 
The gallery was not monitoring for changes to the sorting preference. It
needed a trigger to sort the list anew. This was removed in
d181dba6da9a35b7b. 'Update goals' can stay, it is the fetch goals that
was to be reduced. Furthermore, since gallery's last updated is meant to
indicate how fresh the data is (when they were last fetched), gallery's
setting this to now where it did does not necessarily coincide with
fetching the goals anew. Instead, the goal manager already provides this
nugget.


*For UI changes including screenshots of before and after is great.*

## Validation
Ran app in simulator. 
gallery -> settings -> sort pref, picked another, and back to gallery
from settings
Noticed sort had taken effect.
Also noticed last updated seems much more up-to-date and less out of
sync, as reported in #588.