Do not animate gallery updates (#634)
## Summary
Unfortunately `NSFetchedResultsController` does not provide a minimal
diff when there are core data updates, instead it provides a snapshot
which removes and re-adds all items. When combined with animating
changes, this causes a flicker.

Turn off animation so updates are smooth.

## Validation
Ran on device, pulled to refresh on the gallery view, and observed there
was not a flicker when data updated.