goal not opened when tapping spotlight result (#628)
## Summary
Tapping a spotlight result should open not only the beeminder app but
also the corresponding goal in the app.

## Validation
Searched for the name of a goal. Tapped the Spotlight result
corresponding to beeminder. Noted app was then opened to the goal.


This may have worked before and if so, my guess is that it broke with
Commit 758d7fe.
Fixes #626