Use incremental goal update (#635)
## Summary
Previously the app would always fetch and update all goals on refresh.
This was inefficient both in transfer time and amount of work the server
needed to do, leading to slow updates for users with many goals. Now
instead we ask the server only for updated goals, leading to a smaller
diff.

## Validation
* Launched the app with a test account, loaded a goal, deleted it,
confirmed it was gone
* Added unit tests to cover goal creation and deletion