Update freshness based on user last updated (#636)
## Summary
We were fetching the last updated time for goals from a variable only
stored in memory. This meant when launching the app it would show as
2024 years ago. Now we always update user every time we update goals,
instead fetch the version we keep persisted on the user object.

We also replace all usages of refreshUser with refreshGoals so the two
are always in sync.

## Validation
Hard quit the app. Launched and observed an appropriate value from a few
seconds ago.