alerts not shown when app in foreground (#646)
## Summary
When the device receives a remote notification to display an alert,
either
a) the app is not in the foreground and the alert is shown in the
Notification Center or
b) the app is in the foreground and thus far no alert was shown at all -
neither in the app nor in iOS's Notification Center.


*For UI changes including screenshots of before and after is great.*

## Validation
Dropped a particularly formatted .apns json file onto the simulator.


Perhaps this fixes #460.