sort by recent data and by pledge like website (#648)
## Summary
When sorting the gallery by pledge or by recent data, the smallest
pledges or least recently update goals were shown first. This is in
contrast with what the website does and what the app previously did. The
app currently does not provide for easily flipping the direction of
these sorting tactics.

## Validation
Compared the gallery's order on the website with the app.


Bug was introduced in
[#607](https://github.com/beeminder/BeeSwift/pull/607/files#diff-a606a3bd59aa90e96475a1104ee2119968cb34b90d9fb33afd497f204dcb03e7L357-L363)
where the two previously "reversed" criteria were changed to ascending.