goal not opened when app is opened with url beeminder://?slug=goalname (#643)
## Summary
Opening a link in the format `beeminder://?slug=goalname` previously
opened not only the beeminder app but also the corresponding goal in the
app.

## Validation
Created text with the url and tapped it such that the app opened.
Without the change the app is opened. With the change, the app is opened
and the goal is opened within the app.


This worked before. My guess is that it broke with the transition to
scene delegate, Commit 758d7fe.

Fixes #642.