Search goals also in title (#632)
## Summary
In #631 it was requested that the search works should be extended to
search also in goals' descriptions (here `title`s).

*For UI changes including screenshots of before and after is great.*
https://forum.beeminder.com/t/ios-ux-comments/12243

## Validation
Ran app in simulator. Searched goals for things appearing in titles and
not in goal names. Searched goals for things appearing in names and not
titles. Noticed (expected) results both times.

Fixes #631