Make freshness indicator a bit fuzzy (#659)
## Summary
The freshness indicator only updates from time to time, so showing a
precise number of seconds is a bit misleading. It also leads to thrash
when we update data a bunch of times in a row, each showing a slightly
different value. Instead show an approximation of how long since last
update.

## Validation
Loaded app on device, watched time slowly increase