Make gallery goal cell's thumbnail image border color match safesum (#585)
## Summary
safesum color and thumbnail border were both too similar and different
on light for orange

~~This implementation uses the standard graph image in place of the
thumbnail on the gallery, drawing itself a border around the image. This
has the added benefit of the graph being loaded already when the user
navigates to the corresponding goal screen. It has the downside of
grabbing the larger files (full graph image rather than smaller
thumbnail). Also, the thumbnail image, designed for use on the website,
contained just the graph without its x and y axes' labels and values.~~

## Validation
viewed gallery in simulator


Fixes #573