Support sandbox push notifications for local development (#660)
## Summary
Local and simulator builds use a different push notification server than
production and testflight builds. Tell the beeminder server what type of
build we are so it can use the right apns server.


## Validation
None. Looks like something is currently up with the server side part of
this, but I want to confirm testflight does the expected thing.