open goal corresponding to goal of tapped notification (#645)
This is the app side. Rather than parsing the alert title for the slug
as was done in the [previous PR](#127), this one works with the
notification that provides the slug directly as part of the
user info.

Fixes #41

## Summary
Open the goal corresponding to the tapped notification.

## Validation
Testing by dropping a .apns file into the simulator.