A coordinator handles the app's navigation (#626)
## Summary
Navigation was somewhat scattered throughout the app and the gallery had
a central role. This merge request introduces a class for handling the
app's navigation. Eventually it could also help with scenarios like
starting the app from cold start with the running timer for a particular
goal shown.

*For UI changes including screenshots of before and after is great.*

## Validation
Running the app in the simulator.
Navigating around:
- signing in
- signing out
- viewing gallery
- opening goal from gallery
- opening edit datapoint
- canceling edit datapoint
- opening settings and its children
- opening goal from spotlight result tap
- opening app from tapping notification

Fixes #622


Other Locations:
- [ ] EditDatapointVC dismisses directly upon deleting a datapoint
- [ ] EditDatapointVC dismisses directly upon successfully updating a
datapoint
- [ ] AlertVC are still handled around where they are displayed
- [ ] GoalVC shows the SafariVC
- [ ] TimerVC dismisses itself (indirectly)