Show gallery behind keyboard and fix render jump (#682)
## Summary
Previously the gallery view
* Didn't show goals behind the keyboard (invisible prior to iOS26 but
now the keyboard is meant to be glass)
* Had an awkward render jump when showing/hiding the keyboard.

This addresses this by moving from the theoretically more elegant but in
practice problematic solution of using constraints, to instead setting a
contentInset based on the keyboard appearing / hiding.

## Validation
Loaded the app on device. Showed and hid the keyboard a few times.
Checked the animation looked right. Checked it was possible to scroll to
the very end of the list.