Add dietary fiber HealthKit metric (#689)
Adds dietary fiber as a new Apple Health nutrition metric (following the
existing pattern for other nutrition metrics).

Testing: Tested in simulator. Might be worth testing on a physical
device first. (I'm not yet able to do that (sorry!) but will do so in
the future, after I have the requestSigningKey and all that.)